EIP1-11125: rename openapi specification file and objects defined within it for better naming clarity especially when shared with postal and proxy APIs
